LPD-102356 Move ScopeUtil to the export-import API module
