+++ modules/apps/export-import/export-import-api/src/main/java/com/liferay/exportimport/util/ScopeUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.web.internal.util;
+package com.liferay.exportimport.util;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Group;
