+++ modules/apps/export-import/export-import-api/src/main/resources/com/liferay/exportimport/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
