+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/internal/display/context/ExportImportProcessesDisplayContext.java
@@ -6,7 +6,7 @@
 package com.liferay.exportimport.web.internal.display.context;
 
 import com.liferay.exportimport.web.internal.constants.ExportImportFDSNames;
-import com.liferay.exportimport.web.internal.util.ScopeUtil;
+import com.liferay.exportimport.util.ScopeUtil;
 import com.liferay.frontend.data.set.model.FDSActionDropdownItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenuBuilder;
